implementing user-round-team-city relation

+++ src/main/java/com/around/aroundcore/web/dtos/GameUserDTO.java
@@ -3,8 +3,6 @@ package com.around.aroundcore.web.dtos;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
-import java.util.Objects;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -24,6 +22,6 @@ public class GameUserDTO {
     private Integer coins;
     @Schema(description = "Is user email verified")
     private Boolean verified;
-    private Integer captured_chunks;
+    private Long captured_chunks;
 }
